+++ plugins/decap_cms.ts
@@ -126,10 +126,10 @@ export default function (userOptions?: Options) {
   };
 }
 
-/** Extends PageData interface */
+/** Extends Data interface */
 declare global {
   namespace Lume {
-    export interface PageData {
+    export interface Data {
       /**
        * Decap CMS configuration
        * @see https://lume.land/plugins/decap_cms/