+++ plugins/pug.ts
@@ -152,10 +152,10 @@ export default function (userOptions?: Options) {
   };
 }
 
-/** Extends PageHelpers interface */
+/** Extends Helpers interface */
 declare global {
   namespace Lume {
-    export interface PageHelpers {
+    export interface Helpers {
       /** @see https://lume.land/plugins/pug/ */
       pug: (string: string, data?: Record<string, unknown>) => string;
     }