+++ plugins/search.ts
@@ -21,10 +21,10 @@ export default function (userOptions?: Options) {
   };
 }
 
-/** Extends PageData interface */
+/** Extends Data interface */
 declare global {
   namespace Lume {
-    export interface PageData {
+    export interface Data {
       /**
        * The searcher helper
        * @see https://lume.land/plugins/search/