+++ plugins/terser.ts
@@ -76,10 +76,10 @@ export default function (userOptions?: Options) {
   };
 }
 
-/** Extends PageHelpers interface */
+/** Extends Helpers interface */
 declare global {
   namespace Lume {
-    export interface PageHelpers {
+    export interface Helpers {
       /** @see https://lume.land/plugins/terser/#the-terser-filter */
       terser: (code: string) => Promise<string | undefined>;
     }