+++ plugins/url.ts
@@ -41,10 +41,10 @@ export default function (userOptions?: Options) {
   };
 }
 
-/** Extends PageHelpers interface */
+/** Extends Helpers interface */
 declare global {
   namespace Lume {
-    export interface PageHelpers {
+    export interface Helpers {
       /** @see https://lume.land/plugins/url/#url-filter */
       url: (path: string, absolute?: boolean) => string;