+++ types.ts
@@ -1,5 +1,5 @@
 import type { Engine, Helper } from "./core/renderer.ts";
-import type { Data, Page } from "./core/file.ts";
+import type { Data as PageData, Page } from "./core/file.ts";
 import type { default as Site, Plugin } from "./core/site.ts";
 import type { Archetype } from "./cli/create.ts";
 import type { Middleware, RequestHandler } from "./core/server.ts";
@@ -16,16 +16,14 @@ declare global {
       Site,
     };
 
-    /** The data of a page */
-    // deno-lint-ignore no-explicit-any
-    export interface PageData<Type extends Record<any, any> = any>
-      extends Data, Type {
-      /** The title of the page */
-      title?: string;
+    /** The page data */
+    export interface Data extends PageData {
+      // deno-lint-ignore no-explicit-any
+      [index: string]: any;
     }
 
     /** The page helpers */
-    export interface PageHelpers {
+    export interface Helpers {
       [key: string]: Helper;
     }
   }